chore: update to storybook 7
