+++ ui/package.json
@@ -15,8 +15,8 @@
     "test": "jest --watch",
     "test:ci": "jest --ci",
     "test:coverage": "jest --coverage",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook -o .storybook-static"
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "dependencies": {
     "@chakra-ui/icons": "2.0.18",
@@ -65,24 +65,23 @@
     "zod-formik-adapter": "1.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.12",
+    "@babel/core": "7.21.4",
     "@jest/globals": "29.5.0",
-    "@storybook/addon-actions": "^6.5.15",
-    "@storybook/addon-essentials": "^6.5.15",
-    "@storybook/addon-interactions": "^6.5.15",
-    "@storybook/addon-links": "^6.5.15",
-    "@storybook/builder-webpack5": "^6.5.15",
-    "@storybook/manager-webpack5": "^6.5.15",
-    "@storybook/react": "^6.5.15",
-    "@storybook/testing-library": "^0.0.13",
-    "@testing-library/dom": "8.11.1",
+    "@storybook/addon-essentials": "^7.0.5",
+    "@storybook/addon-interactions": "^7.0.5",
+    "@storybook/addon-links": "^7.0.5",
+    "@storybook/blocks": "^7.0.5",
+    "@storybook/nextjs": "^7.0.5",
+    "@storybook/react": "^7.0.5",
+    "@storybook/testing-library": "0.1.0",
+    "@testing-library/dom": "9.2.0",
     "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "13.1.1",
-    "@testing-library/react-hooks": "7.0.2",
-    "@testing-library/user-event": "13.3.0",
-    "@types/react": "18.0.28",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/user-event": "14.4.3",
+    "@types/react": "18.0.35",
     "@typescript-eslint/eslint-plugin": "5.58.0",
-    "babel-loader": "^8.3.0",
+    "babel-loader": "9.1.2",
     "cross-env": "7.0.3",
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
@@ -91,8 +90,9 @@
     "jest": "28.0.3",
     "jest-environment-jsdom": "28.0.2",
     "prettier": "2.8.7",
-    "storybook-addon-mock": "3.2.0",
-    "storybook-addon-next": "1.7.3",
+    "storybook": "^7.0.5",
+    "storybook-addon-mock": "4.0.0",
+    "storybook-addon-next": "1.8.0",
     "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
     "typescript": "5.0.4"
