+++ ui/tsconfig.json
@@ -10,6 +10,8 @@
     "baseUrl": ".",
     "skipLibCheck": true,
     "strict": false,
+    "strictNullChecks": true,
+    "strictBindCallApply": true,
     "forceConsistentCasingInFileNames": true,
     "noEmit": true,
     "incremental": true,
@@ -29,7 +31,6 @@
         "name": "next"
       }
     ],
-    "strictNullChecks": true
   },
   "include": [
     "next-env.d.ts",
